Adjust metadata following move to Privacy CG
This moves the spec from an "Unofficial Proposal Draft" to a "CG Draft"
following the group's adoption of the spec, adjusts a couple of URLs, and
creates the missing `w3c.json` file.

The boilerplate text in Bikeshed seems to require a LICENSE macro text. This
sets it to the value found in other Privacy CG proposals.